refactor: Re-arranged the files and moved atoms, molecules and organisms folder out of components - leaving only child components in the components folder - PROJ-2704
