+++ packages/web/src/molecules/AppInput.tsx
@@ -25,9 +25,11 @@ const AppInput = React.forwardRef<HTMLInputElement, AppInputProps>(
     return (
       <div className="flex flex-col relative">
         <div
-          className={`flex align-baseline mb-lg pr-md  rounded-[4px] ${isFocused && !error ? 'border-b-2 border-solid !border-light-edge-accent-strong dark:border-dark-edge-accent-strong rounded-b-[0px] transition-all duration-400' : ''} bg-light-surface-gray ${error ? 'border-b-2 border-solid border-light-type-error dark:border-dark-type-error' : ''}`}>
+          className={`flex align-baseline mb-lg pr-md rounded-[4px] ${isFocused && !error ? 'border-b-2 border-solid !border-light-edge-accent-strong dark:border-dark-edge-accent-strong rounded-b-[0px] transition-all duration-400' : ''} bg-light-surface-gray ${error ? 'border-b-2 rounded-b-[0px] border-solid border-light-type-error dark:border-dark-type-error' : ''}`}>
           <input
-            className={`form-control inputFocus2`}
+            className={`
+            bg-light-surface-gray dark:bg-dark-surface-gray !text-light-type-gray dark:text-dark-type-gray outline-none focus:outline-none focus:z-10 appearance-none w-full min-w-[415px] h-[56px] px-md pb-lg !pt-2xl border-0 focus:ring-0 text-sm-head
+              inputFocus2`}
             placeholder={placeholder}
             value={value}
             onFocus={() => setIsFocused(true)}
@@ -39,7 +41,19 @@ const AppInput = React.forwardRef<HTMLInputElement, AppInputProps>(
           />
           <label
             className={`absolute inline-flex left-[11px] top-md transition-all duration-200 pt-xs text-sm-title
-                ${value || isFocused || placeholder ? 'focusLabel' : '!text-light-type-gray-placeholder dark:text-dark-type-gray-placeholder '}`}>
+              peer-focus:transform 
+              peer-focus:origin-top-left 
+              peer-focus:translate-x-[0px] 
+              peer-focus:translate-y-[-12px] 
+              peer-focus:scale-100 
+              peer-focus:ease-[cubic-bezier(0,_0,_0.2,_1)]
+              peer-focus:text-xs-body 
+              peer-focus:z-10
+              ${
+                value || isFocused || placeholder
+                  ? 'transform origin-top-left translate-x-[1px] translate-y-[-12px] scale-100 transition duration-200 ease-[cubic-bezier(0,_0,_0.2,_1)] text-xs-body z-10 !text-light-type-gray-muted dark:text-dark-type-gray-muted '
+                  : '!text-light-type-gray-placeholder dark:text-dark-type-gray-placeholder '
+              }`}>
             {label}
           </label>
           {value && (
@@ -65,5 +79,4 @@ const AppInput = React.forwardRef<HTMLInputElement, AppInputProps>(
     )
   },
 )
-
 export default AppInput
